requirements.txt
anyio
asgiref
certifi
cffi
charset-normalizer
click
colorama
cryptography
diff-match-patch
Django
django-allauth
django-countries
django-import-export
django-solo
et_xmlfile
gunicorn
h11
httptools
idna
openpyxl
packaging
pillow
pycparser
PyJWT
python-dotenv
PyYAML
requests
sniffio
sqlparse
tablib
typing_extensions
tzdata
urllib3
uvicorn
watchfiles
websockets